+++ sphinx/source/changelog.rst
@@ -3,6 +3,9 @@ Changelog
 
 8.4b1
 -----
+* Added the ``Address`` parameter to the :ref:`ref-Game` section (for
+  specifying the format of address fields on disassembly pages and memory map
+  pages, and of the default link text for the :ref:`R` macro)
 * Added support for specifying an :ref:`expand` directive value over multiple
   lines by prefixing the second and subsequent lines with ``+``
 * Changed the default value of the ``DefmSize`` configuration parameter for